# Optimize EventDispatcher performance with Map and Set