--- a/src/core/EventDispatcher.js
+++ b/src/core/EventDispatcher.js
@@ -2,40 +2,58 @@
  * https://github.com/mrdoob/eventdispatcher.js/
  */
 
-class NewEventDispatcher {
-    constructor() {
-        this._listeners = new Map();
-    }
+class EventDispatcher {
 
-    addEventListener(type, listener) {
-        const listeners = this._listeners.get(type) || this._listeners.set(type, new Set()).get(type);
-        listeners.add(listener);
-    }
+	constructor() {
 
-    hasEventListener(type, listener) {
-        const listeners = this._listeners.get(type);
-        return listeners ? listeners.has(listener) : false;
-    }
+		this._listeners = new Map();
 
-    removeEventListener(type, listener) {
-        const listeners = this._listeners.get(type);
-        if (listeners) {
-            listeners.delete(listener);
+	}
 
-            if (listeners.size === 0) {
-                this._listeners.delete(type);
-            }
-        }
-    }
+	addEventListener( type, listener ) {
 
-    dispatchEvent(event) {
-        const listeners = this._listeners.get(event.type);
-        if (listeners) {
-            event.target = this;
-            listeners.forEach(listener => listener.call(this, event));
-            event.target = null;
-        }
-    }
+		const listeners = this._listeners.get( type ) || this._listeners.set( type, new Set() ).get( type );
+		listeners.add( listener );
+
+	}
+
+	hasEventListener( type, listener ) {
+
+		const listeners = this._listeners.get( type );
+		return listeners ? listeners.has( listener ) : false;
+
+	}
+
+	removeEventListener( type, listener ) {
+
+		const listeners = this._listeners.get( type );
+		if ( listeners ) {
+
+			listeners.delete( listener );
+
+			if ( listeners.size === 0 ) {
+
+				this._listeners.delete( type );
+
+			}
+
+		}
+
+	}
+
+	dispatchEvent( event ) {
+
+		const listeners = this._listeners.get( event.type );
+		if ( listeners ) {
+
+			event.target = this;
+			listeners.forEach( listener => listener.call( this, event ) );
+			event.target = null;
+
+		}
+
+	}
+
 }
 
 export { EventDispatcher };