Hey, not sure if these changes are what you're looking for, but here's what I did:

I swapped out plain objects and arrays for `Map` and `Set` in the `EventDispatcher` class to store event listeners. The goal is to boost performance, especially when there are a lot of listeners, since `Map` and `Set` offer O(1) time complexity for adding, removing, and checking items.

### Key Changes:
- **Performance improvement**: Should give a nice performance improvement in `Edge`, and we should see similar improvements in `Chrome` and `Firefox` too.
- **API compatibility**: The update doesn't break the existing `EventDispatcher` API, so no need to change anything on the user side I think.
- **Efficiency improvements**: This should help reduce overhead and make event dispatching faster, which is especially useful in performance-heavy scenarios like 3D games.
- **Cross-browser compatibility**: It's all compatible with the MDN baseline and works across modern browsers.
- **Benchmarks**: I ran some benchmarks with `performance.now()` in the browser and saw good improvements in `Edge`. `Chrome` should have similar results.

Let me know what you think!