@mrdoob 2024-11-19
> * **Benchmarks**: I ran some benchmarks with `performance.now()` in the browser and saw good improvements in `Edge`. `Chrome` should have similar results.

Please, share links to these benchmarks 🙏