@linbingquan 2024-11-19
```ts
import { EventDispatcher as OldEventDispatcher } from "./old.js";
import { EventDispatcher as NewEventDispatcher } from "./new.js";

Deno.bench({
  name: "new",
  fn() {
    const dispatcher = new NewEventDispatcher();
    dispatcher.addEventListener("event", () => {});
    dispatcher.dispatchEvent("event");
  },
});

Deno.bench({
  name: "old",
  fn() {
    const dispatcher = new OldEventDispatcher();
    dispatcher.addEventListener("event", () => {});
    dispatcher.dispatchEvent("event");
  },
});
```

```
benchmark   time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------- ----------------------------- --------------------- --------------------------
old                 47.3 ns    21,140,000 ( 39.5 ns … 247.7 ns)  48.9 ns  92.1 ns  97.4 ns
new                115.2 ns     8,677,000 ( 99.8 ns … 629.5 ns) 108.6 ns 552.0 ns 607.8 ns

summary
  old
     2.44x faster than new
```