@linbingquan 2024-11-19
```ts
import { EventDispatcher as OldEventDispatcher } from "./old.js";
import { EventDispatcher as NewEventDispatcher } from "./new.js";

Deno.bench({
  name: "old",
  baseline: true,
  fn() {
    const dispatcher = new OldEventDispatcher();
    dispatcher.addEventListener("event", () => {});
    dispatcher.dispatchEvent("event");
  },
});

Deno.bench({
  name: "new",
  fn() {
    const dispatcher = new NewEventDispatcher();
    dispatcher.addEventListener("event", () => {});
    dispatcher.dispatchEvent("event");
  },
});
```

```
benchmark   time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------- ----------------------------- --------------------- --------------------------
old                 48.1 ns    20,790,000 ( 39.1 ns … 201.3 ns)  49.4 ns  95.1 ns 109.9 ns
new                116.6 ns     8,574,000 (102.7 ns … 297.5 ns) 119.9 ns 191.1 ns 218.7 ns

summary
  old
     2.42x faster than new
```