@sunag 2024-11-19
> old 2.42x faster than new

It makes sense to me. Usually primitive values ​​are faster, this should also apply to `for` instead of `forEach`.