@gkjohnson 2024-11-19
> ```js
> dispatcher.dispatchEvent("event");
> ```

Just to make sure - this benchmark isn't using `dispatchEvent` properly. The function takes an object like so:

> ```js
> dispatcher.dispatchEvent( { type: 'event' } );
> ```

It may still be faster but this "dispatch" call isn't actually doing or testing anything right now. I would also benchmark `dispatchEvent` and `addEventListener` separately since `dispatchEvent` will get called significantly more frequently than `addEventListener`.