@eXponenta 2024-11-22
Any iterators that use `Symbol.iterator` is slower about 2 times relative indexed for loop. 

Map and Set stored as linked hash tree and can't be indexed and because of this uses `Symbol.iterator` for it.

Iteration more important than add/remove listeners, because events can be fired every frame, but listeners assigned only in app start or when something updated by user input.

Also classic events tree about 10-20 listeners, slice over 10-20 is highly optimised operation.

Also Map and Set by designed for frequent update by unknown key set.

Events map is KNOWN keyset. Because we not set random key as event type every update and listeners map will be high optimised type-shape with strictly known properties.

Classic event name in core: "update".
This mean that we will be created a lot of  small objects with references to single shape map inside JS runtime with property name "update".
[How this works] (https://mathiasbynens.be/notes/shapes-ics )


Conclusion: Map and Set not good idea for usage for EventDispatcher due to performance degradations in basic usage.