@linbingquan 2024-12-23
I borrowed from the existing library and I wrote `Emitter.ts` and found these:

With the following changes, dispatchEvent should be faster than it is now.

Test cases can be less rigorous and contain errors.

1.  `dispatchEvent( event )` change to `dispatchEvent(type, ...args)`
2.  remove code `event.target = this` and `event.target = null` 
3.  use `apply` replace `call`

```sh
benchmark            time/iter (avg)        iter/s      (min … max)           p75      p99     p995
-------------------- ----------------------------- --------------------- --------------------------
Emitter                      18.7 ns    53,360,000 ( 15.3 ns … 138.2 ns)  18.9 ns  36.4 ns  55.3 ns
NewEventDispatcher           18.8 ns    53,110,000 ( 14.6 ns … 157.2 ns)  18.4 ns  73.1 ns  82.2 ns
EventDispatcher              30.8 ns    32,450,000 ( 24.0 ns … 244.5 ns)  33.5 ns  58.2 ns 118.6 ns
mitt                         38.4 ns    26,030,000 ( 32.2 ns … 142.8 ns)  41.3 ns  69.3 ns 100.2 ns
onfire.js                    41.3 ns    24,230,000 ( 30.5 ns … 249.9 ns)  44.2 ns  81.0 ns 120.2 ns

summary
  Emitter
     1.00x faster than NewEventDispatcher
     1.64x faster than EventDispatcher
     2.05x faster than mitt
     2.20x faster than onfire.js
```

If these changes are implemented, they should involve quite a few code changes...
We may must to explicitly pass the target like: `dispatchEvent(type, {type, target: this})`